requirements.txt
streamlit
qdrant-client
sentence-transformers
llama-cpp-python
pymupdf
torch
